+++ packages/uniforms-antd/src/NestField.js
@@ -40,4 +40,4 @@ const Nest = ({
     </div>
 ;
 
-export default connectField(Nest, {includeInChain: false});
+export default connectField(Nest, {ensureValue: false, includeInChain: false});