+++ packages/uniforms-bootstrap4/src/ListField.js
@@ -67,4 +67,4 @@ const List = ({
     </div>
 ;
 
-export default connectField(List, {includeInChain: false});
+export default connectField(List, {ensureValue: true, includeInChain: false});