+++ packages/uniforms-material/src/ListField.js
@@ -43,4 +43,4 @@ const List = ({
     </ListMaterial>
 ;
 
-export default connectField(List, {includeInChain: false});
+export default connectField(List, {ensureValue: true, includeInChain: false});