+++ packages/uniforms-material/src/NestField.js
@@ -29,4 +29,4 @@ const Nest = ({
     </div>
 ;
 
-export default connectField(Nest, {includeInChain: false});
+export default connectField(Nest, {ensureValue: false, includeInChain: false});