+++ packages/uniforms-semantic/src/NestField.js
@@ -48,4 +48,4 @@ const Nest = ({
 
 Nest.defaultProps = {grouped: true};
 
-export default connectField(Nest, {includeInChain: false});
+export default connectField(Nest, {ensureValue: false, includeInChain: false});