+++ packages/uniforms-unstyled/src/NestField.js
@@ -31,4 +31,4 @@ const Nest = ({
     </div>
 ;
 
-export default connectField(Nest, {includeInChain: false});
+export default connectField(Nest, {ensureValue: false, includeInChain: false});